+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Please choose versions by [Semantic Versioning](http://semver.org/).
 * MINOR version when you add functionality in a backwards-compatible manner, and
 * PATCH version when you make backwards-compatible bug fixes.
 
-## Unreleased
+## v0.12.2
 
 - Auto-generate changelog entries from git commits when ## Unreleased is missing
 - Check commits since last tag to detect unreleased changes